Refactor SystemServicesContext
